requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy~=1.26.4
pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
albumentations>=1.3.0
tensorboard>=2.13.0
click>=8.1.0
loguru>=0.7.0
hydra-core>=1.3.0
omegaconf>=2.3.0
